Move Zuora models to support-models
