+++ src/main/scala/com/gu/support/workers/encoding/Encoding.scala
@@ -3,10 +3,11 @@ package com.gu.support.workers.encoding
 import java.io.{InputStream, OutputStream}
 import java.util.Base64
 
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.encoding.Encryption._
 import com.gu.support.workers.encoding.Wrapper._
 import com.gu.support.workers.model.{ExecutionError, RequestInfo}
-import com.gu.zuora.encoding.CustomCodecs.{jsonWrapperDecoder, jsonWrapperEncoder}
+import io.circe.generic.auto._
 
 import scala.util.Try
