--- src/main/scala/com/gu/support/workers/encoding/Helpers.scala
@@ -1,15 +1,0 @@
-package com.gu.support.workers.encoding
-
-import com.gu.zuora.encoding.CapitalizationEncoder._
-import io.circe.generic.decoding.DerivedDecoder
-import io.circe.generic.encoding.DerivedObjectEncoder
-import io.circe.generic.semiauto._
-import shapeless.Lazy
-
-object Helpers {
-  def deriveCodec[A](implicit decode: Lazy[DerivedDecoder[A]], encode: Lazy[DerivedObjectEncoder[A]]): Codec[A] =
-    new Codec(deriveEncoder, deriveDecoder)
-
-  def capitalizingCodec[A](implicit decode: Lazy[DerivedDecoder[A]], encode: Lazy[DerivedObjectEncoder[A]]): Codec[A] =
-    new Codec(capitalizingEncoder, decapitalizingDecoder)
-}
