+++ src/main/scala/com/gu/support/workers/encoding/StateDecoder.scala
@@ -4,9 +4,10 @@ import java.util.Base64
 
 import com.gu.config.Configuration
 import com.gu.support.config.AwsConfig
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.model.JsonWrapper
-import com.gu.zuora.encoding.CustomCodecs.jsonWrapperDecoder
 import com.typesafe.scalalogging.LazyLogging
+import io.circe.generic.auto._
 import io.circe.parser._
 
 import scala.util.Try
