+++ src/main/scala/com/gu/support/workers/exceptions/ErrorHandler.scala
@@ -8,7 +8,7 @@ import com.gu.paypal.PayPalError
 import com.gu.salesforce.Salesforce.SalesforceErrorResponse
 import com.gu.stripe.Stripe
 import com.gu.support.workers.exceptions.RetryImplicits._
-import com.gu.zuora.model.response.ZuoraErrorResponse
+import com.gu.support.zuora.api.response.ZuoraErrorResponse
 /**
  * Maps exceptions from the application to either fatal or non fatal exceptions
  * based on whether we think retrying them has a chance of succeeding
