+++ src/main/scala/com/gu/support/workers/lambdas/FailureHandler.scala
@@ -5,13 +5,14 @@ import com.gu.emailservices.{EmailService, FailedContributionEmailFields, Failed
 import com.gu.helpers.FutureExtensions._
 import com.gu.monitoring.SafeLogger
 import com.gu.stripe.Stripe.StripeError
-import com.gu.support.workers.encoding.ErrorJson
-import com.gu.support.workers.encoding.StateCodecs._
+import com.gu.support.encoding.CustomCodecs._
+import com.gu.support.encoding.ErrorJson
 import com.gu.support.workers.model.CheckoutFailureReasons._
 import com.gu.support.workers.model.states.{CheckoutFailureState, FailureHandlerState}
 import com.gu.support.workers.model.{Contribution, DigitalPack, ExecutionError, RequestInfo}
-import com.gu.zuora.model.response.{ZuoraError, ZuoraErrorResponse}
+import com.gu.support.zuora.api.response.{ZuoraError, ZuoraErrorResponse}
 import io.circe.Decoder
+import io.circe.generic.auto._
 import io.circe.parser.decode
 
 import scala.concurrent.ExecutionContext.Implicits.global
