+++ src/main/scala/com/gu/support/workers/lambdas/SendThankYouEmail.scala
@@ -6,12 +6,12 @@ import com.gu.emailservices.{ContributionEmailFields, DigitalPackEmailFields, Em
 import com.gu.monitoring.SafeLogger
 import com.gu.salesforce.Salesforce.SfContactId
 import com.gu.services.{ServiceProvider, Services}
-import com.gu.support.workers.encoding.StateCodecs._
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.model.states.SendThankYouEmailState
 import com.gu.support.workers.model.{Contribution, DigitalPack, DirectDebitPaymentMethod, RequestInfo}
 import com.gu.threadpools.CustomPool.executionContext
 import com.gu.zuora.ZuoraService
-import com.gu.zuora.encoding.CustomCodecs._
+import io.circe.generic.auto._
 import org.joda.time.DateTime
 
 import scala.concurrent.Future
