--- src/main/scala/com/gu/zuora/model/ContactDetails.scala
@@ -1,21 +1,0 @@
-package com.gu.zuora.model
-
-import com.gu.i18n.Country
-import com.gu.support.workers.encoding.Codec
-import com.gu.support.workers.encoding.Helpers.capitalizingCodec
-import com.gu.zuora.encoding.CustomCodecs._
-
-object ContactDetails {
-  implicit val codec: Codec[ContactDetails] = capitalizingCodec
-}
-case class ContactDetails(
-  firstName: String,
-  lastName: String,
-  workEmail: String,
-  country: Country,
-  address1: Option[String] = None,
-  address2: Option[String] = None,
-  city: Option[String] = None,
-  postalCode: Option[String] = None,
-  state: Option[String] = None
-)
