--- src/main/scala/com/gu/zuora/model/SubscribeOptions.scala
@@ -1,9 +1,0 @@
-package com.gu.zuora.model
-
-import com.gu.support.workers.encoding.Codec
-import com.gu.support.workers.encoding.Helpers.capitalizingCodec
-
-object SubscribeOptions {
-  implicit val codec: Codec[SubscribeOptions] = capitalizingCodec
-}
-case class SubscribeOptions(generateInvoice: Boolean = true, processPayments: Boolean = true)
