--- src/main/scala/com/gu/zuora/model/package.scala
@@ -1,5 +1,0 @@
-package com.gu.zuora
-
-package object model {
-  type RatePlanId = String
-}
