--- src/test/resources/logback-test.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- Turn off logging during tests -->
-<configuration />
