+++ src/test/scala/com/gu/support/workers/Fixtures.scala
@@ -5,10 +5,11 @@ import java.io.ByteArrayInputStream
 import com.gu.i18n.Currency
 import com.gu.i18n.Currency.GBP
 import com.gu.salesforce.Fixtures.{email, idId}
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.encoding.Conversions.StringInputStreamConversions
 import com.gu.support.workers.encoding.Wrapper
 import com.gu.support.workers.model.{BillingPeriod, Monthly, RequestInfo}
-import com.gu.zuora.encoding.CustomCodecs.jsonWrapperEncoder
+import io.circe.generic.auto._
 import io.circe.syntax._
 
 //noinspection TypeAnnotation
