+++ src/test/scala/com/gu/support/workers/integration/CreateSalesforceContactSpec.scala
@@ -3,14 +3,15 @@ package com.gu.support.workers.integration
 import java.io.ByteArrayOutputStream
 
 import com.gu.salesforce.Fixtures.salesforceId
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.Fixtures.{createSalesForceContactJson, wrapFixture}
 import com.gu.support.workers.LambdaSpec
 import com.gu.support.workers.encoding.Conversions.FromOutputStream
 import com.gu.support.workers.encoding.Encoding
-import com.gu.support.workers.encoding.StateCodecs._
 import com.gu.support.workers.lambdas.CreateSalesforceContact
 import com.gu.support.workers.model.states.CreateZuoraSubscriptionState
 import com.gu.test.tags.annotations.IntegrationTest
+import io.circe.generic.auto._
 
 @IntegrationTest
 class CreateSalesforceContactSpec extends LambdaSpec {
