+++ src/test/scala/com/gu/support/workers/integration/CreateZuoraSubscriptionSpec.scala
@@ -4,19 +4,20 @@ import java.io.ByteArrayOutputStream
 
 import com.gu.config.Configuration.zuoraConfigProvider
 import com.gu.okhttp.RequestRunners.configurableFutureRunner
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.Fixtures._
 import com.gu.support.workers.LambdaSpec
 import com.gu.support.workers.encoding.Conversions.FromOutputStream
 import com.gu.support.workers.encoding.Encoding
-import com.gu.support.workers.encoding.StateCodecs._
 import com.gu.support.workers.errors.MockServicesCreator
 import com.gu.support.workers.lambdas.{CreateZuoraSubscription, IdentityId}
 import com.gu.support.workers.model.states.SendThankYouEmailState
 import com.gu.support.workers.model.{Annual, Monthly}
+import com.gu.support.zuora.api.SubscribeRequest
+import com.gu.support.zuora.api.response.ZuoraAccountNumber
 import com.gu.test.tags.annotations.IntegrationTest
-import com.gu.zuora.GetAccountForIdentity.ZuoraAccountNumber
 import com.gu.zuora.ZuoraService
-import com.gu.zuora.model.SubscribeRequest
+import io.circe.generic.auto._
 import org.mockito.Matchers.any
 import org.mockito.Mockito.when
 import org.mockito.invocation.InvocationOnMock
