+++ src/test/scala/com/gu/support/workers/integration/SendThankYouEmailSpec.scala
@@ -8,6 +8,7 @@ import com.gu.i18n.Country.UK
 import com.gu.i18n.Currency
 import com.gu.i18n.Currency.GBP
 import com.gu.salesforce.Salesforce.SfContactId
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.Fixtures.{thankYouEmailJson, wrapFixture}
 import com.gu.support.workers.LambdaSpec
 import com.gu.support.workers.encoding.Conversions.FromOutputStream
@@ -16,8 +17,8 @@ import com.gu.support.workers.lambdas.SendThankYouEmail
 import com.gu.support.workers.model.{DirectDebitPaymentMethod, Monthly, User}
 import com.gu.test.tags.annotations.IntegrationTest
 import com.gu.threadpools.CustomPool.executionContext
-import com.gu.zuora.encoding.CustomCodecs._
 import io.circe.Json
+import io.circe.generic.auto._
 import io.circe.parser._
 import org.joda.time.DateTime
