+++ src/test/scala/com/gu/support/workers/lambdas/CreatePaymentMethodSpec.scala
@@ -10,16 +10,16 @@ import com.gu.okhttp.RequestRunners.configurableFutureRunner
 import com.gu.services.{ServiceProvider, Services}
 import com.gu.stripe.Stripe.{StripeError, StripeList}
 import com.gu.stripe.{Stripe, StripeService}
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.Fixtures.{validBaid, _}
 import com.gu.support.workers.encoding.Conversions.{FromOutputStream, StringInputStreamConversions}
 import com.gu.support.workers.encoding.Encoding
-import com.gu.support.workers.encoding.StateCodecs._
 import com.gu.support.workers.exceptions.RetryNone
 import com.gu.support.workers.model.states.CreateSalesforceContactState
 import com.gu.support.workers.model.{CreditCardReferenceTransaction, PayPalReferenceTransaction, PaymentMethod}
 import com.gu.support.workers.{AsyncLambdaSpec, MockContext}
 import com.gu.test.tags.objects.IntegrationTest
-import com.gu.zuora.encoding.CustomCodecs._
+import io.circe.generic.auto._
 import org.mockito.Matchers._
 import org.mockito.Mockito._
