+++ src/test/scala/com/gu/support/workers/lambdas/CreatePaymentMethodStateDecoderSpec.scala
@@ -1,13 +1,13 @@
 package com.gu.support.workers.lambdas
 
 import com.gu.i18n.Currency.GBP
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.Fixtures.{validBaid, _}
-import com.gu.support.workers.encoding.StateCodecs._
 import com.gu.support.workers.model._
 import com.gu.support.workers.model.states.CreatePaymentMethodState
-import com.gu.zuora.encoding.CustomCodecs._
 import com.typesafe.scalalogging.LazyLogging
 import io.circe.Json
+import io.circe.generic.auto._
 import io.circe.parser._
 import io.circe.syntax._
 import org.scalatest.mockito.MockitoSugar
