+++ src/test/scala/com/gu/support/workers/lambdas/CreateSalesforceContactDecoderSpec.scala
@@ -1,11 +1,11 @@
 package com.gu.support.workers.lambdas
 
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.Fixtures._
-import com.gu.support.workers.encoding.StateCodecs._
 import com.gu.support.workers.model.states.CreateSalesforceContactState
 import com.gu.support.workers.model.{Contribution, PayPalReferenceTransaction, PaymentMethod}
-import com.gu.zuora.encoding.CustomCodecs._
 import com.typesafe.scalalogging.LazyLogging
+import io.circe.generic.auto._
 import io.circe.parser._
 import org.scalatest.mockito.MockitoSugar
 import org.scalatest.{FlatSpec, Matchers}
