+++ src/test/scala/com/gu/zuora/ZuoraSpec.scala
@@ -8,11 +8,10 @@ import com.gu.okhttp.RequestRunners
 import com.gu.support.workers.GetRecurringSubscription
 import com.gu.support.workers.lambdas.IdentityId
 import com.gu.support.workers.model.Monthly
+import com.gu.support.zuora.api.SubscribeRequest
+import com.gu.support.zuora.api.response.{ZuoraAccountNumber, ZuoraErrorResponse}
 import com.gu.test.tags.annotations.IntegrationTest
 import com.gu.zuora.Fixtures._
-import com.gu.zuora.GetAccountForIdentity.ZuoraAccountNumber
-import com.gu.zuora.model.SubscribeRequest
-import com.gu.zuora.model.response.ZuoraErrorResponse
 import org.scalatest.{AsyncFlatSpec, Matchers}
 
 import scala.concurrent.duration._
